Rename acr package to azure
