+++ pkg/credentials/azure/workload_identity_test.go
@@ -1,4 +1,4 @@
-package acr
+package azure
 
 import (
 	"context"
